Finished refactoring (compiles, but buggy)
