+++ sqrl-engines/sqrl-engine-core/src/main/java/com/datasqrl/plan/queries/APISource.java
@@ -5,6 +5,7 @@ import com.datasqrl.canonicalizer.NameCanonicalizer;
 import com.datasqrl.canonicalizer.NamePath;
 import com.datasqrl.module.resolver.ResourceResolver;
 import com.datasqrl.util.FileUtil;
+import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.EqualsAndHashCode.Include;
 import lombok.Value;
@@ -15,8 +16,11 @@ public class APISource {
 
   @Include
   Name name;
-  String path;
   String schemaDefinition;
+
+  public static APISource of(String schemaDefinition) {
+    return new APISource(Name.system("schema"),schemaDefinition);
+  }
 
   @Override
   public String toString() {
@@ -26,7 +30,6 @@ public class APISource {
   public static APISource of(String filename, NameCanonicalizer canonicalizer, ResourceResolver resolver) {
     return new APISource(
         canonicalizer.name(FileUtil.separateExtension(filename).getKey()),
-        filename,
         FileUtil.readFile(resolver.resolveFile(NamePath.of(filename)).get())
     );
   }
