+++ sqrl-planner/sqrl-planner-local/src/main/java/com/datasqrl/graphql/APIConnectorManager.java
@@ -33,44 +33,7 @@ import java.util.stream.Collectors;
 import lombok.Getter;
 import org.apache.calcite.rel.type.RelDataTypeFactory;
 
-@Getter
-public class APIConnectorManager {
-
-  private final Optional<LogEngine> logEngine;
-  private final ErrorCollector errors;
-  private final ModuleLoader moduleLoader;
-  private final RelDataTypeFactory typeFactory;
-
-  private final Map<APIMutation, TableSink> mutations = new HashMap<>();
-
-  private final Map<NamePath, LogModule> modules = new HashMap<>();
-
-  private final Map<APISubscription, TableSource> subscriptions = new HashMap<>();
-
-  private final Map<SQRLTable, Log> exports = new HashMap<>();
-
-  private final List<APIQuery> queries = new ArrayList<>();
-
-
-  @Inject
-  public APIConnectorManager(ExecutionPipeline pipeline,
-      ErrorCollector errors, ModuleLoader moduleLoader, RelDataTypeFactory typeFactory) {
-    this.errors = errors;
-    this.moduleLoader = moduleLoader;
-    this.logEngine = pipeline.getStage(Type.LOG).map(stage -> (LogEngine) stage.getEngine());
-    this.typeFactory = typeFactory;
-  }
-
-  private APIConnectorManager() {
-    this.errors = ErrorCollector.root();
-    this.moduleLoader = null;
-    this.logEngine = Optional.empty();
-    this.typeFactory = null;
-  }
-
-  public static APIConnectorManager testInstance() {
-    return new APIConnectorManager();
-  }
+public interface APIConnectorManager {
 
   /**
    * Adds mutation by connecting it to a table source and sink.
@@ -78,103 +41,17 @@ public class APIConnectorManager {
    *
    * @param mutation
    */
-  public void addMutation(APIMutation mutation) {
-    NamePath apiNamePath = apiToModulePath(mutation.getSource());
-    Optional<SqrlModule> module = moduleLoader.getModule(apiNamePath);
-    if (module.isPresent()) {
-      Optional<NamespaceObject> log = module.get().getNamespaceObject(mutation.getName());
-      errors.checkFatal(log.isPresent(), "Could not load mutation endpoint for %s from module %s",
-          mutation, module.get());
-      errors.checkFatal(log.get() instanceof TableSourceSinkNamespaceObject, "Loaded mutation endpoint for %s from module %s is not a source and sink",
-          mutation, module.get());
-      TableSourceSinkNamespaceObject sourceSink = (TableSourceSinkNamespaceObject) log.get();
-      mutations.put(mutation, sourceSink.getSink());
-    } else {
-      //Create module if log engine is set
-      errors.checkFatal(logEngine.isPresent(), "Cannot create mutation %s: Could not load "
-          + "module for %s and no log engine configured", mutation, apiNamePath);
-      LogModule logModule = modules.get(apiNamePath);
-      if (logModule==null) {
-        logModule = new LogModule();
-        modules.put(apiNamePath, logModule);
-      }
-      String logId = getLogId(mutation);
-      Log log = logEngine.get().createLog(logId, mutation.getSchema());
-      logModule.addEntry(mutation.getName(),log);
-      mutations.put(mutation, log.getSink());
-    }
-  }
+  void addMutation(APIMutation mutation);
 
-  public TableSource addSubscription(APISubscription subscription, SQRLTable sqrlTable) {
-    errors.checkFatal(logEngine.isPresent(), "Cannot create subscriptions because no log engine is configured");
-    errors.checkFatal(sqrlTable.getVt().getRoot().getBase().getType()== TableType.STREAM,
-        "Table %s for subscription %s is not a stream table", sqrlTable, subscription);
-    //Check if we already exported it
-    TableSource subscriptionSource;
-    if (exports.containsKey(sqrlTable)) {
-      subscriptionSource = exports.get(sqrlTable).getSource();
-    } else {
-      //otherwise create new log for it
-      String logId = sqrlTable.getVt().getNameId();
-      RelDataType2UTBConverter converter = new RelDataType2UTBConverter(typeFactory, 0,
-          NameCanonicalizer.SYSTEM);
-      UniversalTable schema = converter.convert(sqrlTable.getPath(), sqrlTable.getVt().getRowType(),
-          null);
-      Log log = logEngine.get().createLog(logId, schema);
-      exports.put(sqrlTable, log);
-      subscriptionSource = log.getSource();
-    }
-    subscriptions.put(subscription, subscriptionSource);
-    return subscriptionSource;
-  }
+  TableSource addSubscription(APISubscription subscription, SQRLTable sqrlTable);
 
-  public void addQuery(APIQuery query) {
-    queries.add(query);
-  }
+  void addQuery(APIQuery query);
 
-  public ModuleLoader getAsModuleLoader() {
-    return ModuleLoaderPreloaded.builder().modules(modules).build();
-  }
+  ModuleLoader getAsModuleLoader();
 
-  public List<Log> getLogs() {
-    List<Log> logs = new ArrayList<>();
-    logs.addAll(exports.values());
-    modules.values().stream()
-        .flatMap(logModule -> logModule.entries.values().stream())
-        .forEach(logs::add);
-    return logs;
-  }
+  List<Log> getLogs();
 
-  private NamePath apiToModulePath(APISource source) {
-    return source.getName().toNamePath();
-  }
+  List<APIQuery> getQueries();
 
-  private String getLogId(APIMutation mutation) {
-    return mutation.getSource().getName().getCanonical() + "-" + mutation.getName().getCanonical();
-  }
-
-
-  private class LogModule implements SqrlModule {
-
-    Map<Name, Log> entries = new HashMap<>();
-
-    void addEntry(Name name, Log log) {
-      errors.checkFatal(!entries.containsKey(name),"Log entry of name %s already exists", name);
-      entries.put(name, log);
-    }
-
-    @Override
-    public Optional<NamespaceObject> getNamespaceObject(Name name) {
-      return Optional.ofNullable(entries.get(name)).map(log ->
-          new TableSourceSinkNamespaceObject(log.getSource(), log.getSink()));
-    }
-
-    @Override
-    public List<NamespaceObject> getNamespaceObjects() {
-      return entries.values().stream().map(log ->
-          new TableSourceSinkNamespaceObject(log.getSource(), log.getSink()))
-          .collect(Collectors.toList());
-    }
-  }
-
+  Map<SQRLTable, Log> getExports();
 }
