+++ sqrl-planner/sqrl-planner-local/src/main/java/com/datasqrl/graphql/inference/GraphQLMutationExtraction.java
@@ -28,8 +28,8 @@ public class GraphQLMutationExtraction {
   private final NameCanonicalizer canonicalizer;
 
   @Inject
-  public GraphQLMutationExtraction(SqrlSchema sqrlSchema, NameCanonicalizer nameCanonicalizer) {
-    this.typeFactory = sqrlSchema.getTypeFactory();
+  public GraphQLMutationExtraction(RelDataTypeFactory typeFactory, NameCanonicalizer nameCanonicalizer) {
+    this.typeFactory = typeFactory;
     this.canonicalizer = nameCanonicalizer;
   }
