+++ sqrl-planner/sqrl-planner-local/src/test/java/com/datasqrl/graphql/inference/SchemaInferenceTest.java
@@ -5,14 +5,18 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.datasqrl.canonicalizer.Name;
+import com.datasqrl.graphql.APIConnectorManager;
 import com.datasqrl.graphql.inference.SchemaInferenceModel.InferredSchema;
 import com.datasqrl.loaders.ModuleLoader;
+import com.datasqrl.plan.local.analyze.MockAPIConnectorManager;
+import com.datasqrl.plan.queries.APISource;
 import com.datasqrl.schema.Column;
 import com.datasqrl.schema.SQRLTable;
 import java.util.List;
 import java.util.Optional;
 import org.apache.calcite.jdbc.SqrlSchema;
 import org.apache.calcite.tools.RelBuilder;
+import org.apiguardian.api.API;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -41,16 +45,17 @@ class SchemaInferenceTest {
     /*
      * IMPORT User;
      */
+    APISource gql = APISource.of("type Query { "
+        + "  getUsers: [User] "
+        + "  getUserById(id: String!): User "
+        + "}"
+        + "type User { id: String }");
+
     SchemaInference schemaInference = new SchemaInference(
-        loader, "schema",
-        "type Query { "
-            + "  getUsers: [User] "
-            + "  getUserById(id: String!): User "
-            + "}"
-            + "type User { id: String }",
+        loader, gql,
         schema,
         relBuilder,
-        null
+        null, new MockAPIConnectorManager()
     );
 
     InferredSchema inferredSchema = schemaInference.accept();
@@ -67,22 +72,22 @@ class SchemaInferenceTest {
     Column username = mock(Column.class);
     when(table.getField(Name.system("username"))).thenReturn(Optional.of(username));
 
+
+    APISource gql = APISource.of("type Query { "
+        + "  getUsers: [UserReduced] "
+        + "  getUsersExtended: [UserExtended] "
+        + "}"
+        + "interface User { id: String }"
+        + "type UserReduced implements User { username: String }"
+        + "type UserExtended implements User { id: String, username: String }");
     /*
      * IMPORT User;
      */
     SchemaInference schemaInference = new SchemaInference(
-        loader, "schema",
-        "type Query { "
-            + "  getUsers: [UserReduced] "
-            + "  getUsersExtended: [UserExtended] "
-            + "}"
-            + "interface User { id: String }"
-            + "type UserReduced implements User { username: String }"
-            + "type UserExtended implements User { id: String, username: String }"
-            + "",
+        loader, gql,
         schema,
         relBuilder,
-        null
+        null, new MockAPIConnectorManager()
     );
 
     InferredSchema inferredSchema = schemaInference.accept();
